ts node and break up classes
