+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "tsc && node index.js"
+    "start": "ts-node server.ts",
+    "dev": "ts-node-dev server.ts"
   },
   "author": "",
   "license": "ISC",
@@ -18,6 +19,7 @@
     "@types/d3": "^7.4.0",
     "@types/node": "^20.2.3",
     "@types/ws": "^8.5.4",
+    "ts-node": "^10.9.1",
     "typescript": "^5.0.4"
   }
 }
